+++ services/ui-src/src/components/index.tsx
@@ -36,4 +36,3 @@ export * from "./SupportInfo";
 export * from "./Table";
 export * from "./Upload";
 export * from "./YesNoModalDialog";
-export * from "./AABRate";
