+++ services/ui-src/src/measures/2022/CommonQuestions/OptionalMeasureStrat/context.ts
@@ -20,6 +20,12 @@ interface ContextProps {
   customDenominatorLabel?: string;
   customRateLabel?: string;
   customPrompt?: string;
+  rateCalculation?: (
+    numerator: string,
+    denominator: string,
+    rateMultiplicationValue: number,
+    numbersAfterDecimal: number
+  ) => string;
 }
 
 const PerformanceMeasureContext = createContext<ContextProps>({
