+++ services/ui-src/src/measures/2022/CommonQuestions/OtherPerformanceMeasure/index.tsx
@@ -14,6 +14,12 @@ interface Props {
   allowNumeratorGreaterThanDenominator?: boolean;
   data?: DataDrivenTypes.PerformanceMeasure;
   RateComponent?: RateComp;
+  rateCalc?: (
+    numerator: string,
+    denominator: string,
+    rateMultiplicationValue: number,
+    numbersAfterDecimal: number
+  ) => string;
 }
 
 const stringIsReadOnly = (dataSource: string) => {
@@ -36,6 +42,7 @@ export const OtherPerformanceMeasure = ({
   allowNumeratorGreaterThanDenominator,
   data = {},
   RateComponent = QMR.Rate,
+  rateCalc,
 }: Props) => {
   const register = useCustomRegister<Types.OtherPerformanceMeasure>();
   const { getValues } = useFormContext<Types.OtherPerformanceMeasure>();
@@ -112,6 +119,7 @@ export const OtherPerformanceMeasure = ({
                 allowNumeratorGreaterThanDenominator={
                   allowNumeratorGreaterThanDenominator
                 }
+                rateCalc={rateCalc}
               />
             </CUI.Stack>
           );
